Extract the streaming renderer into its own handler module
`webui-handler/src/lib.rs` had grown to 11,259 lines while carrying the new
streaming wire contract, against the repo's own "split at ~400 lines" rule.
The crate already has ten sibling modules, so `lib.rs` was the outlier.

Move the streaming renderer into `crates/webui-handler/src/streaming/`, split
by concern: transport sink and render entry (`mod`), request-local state
(`state`), cold error constructors (`error`), component-root parity (`root`),
checkpoint inventory and reachability (`inventory`), and checkpoint emission
(`checkpoint`). Each stays under 400 lines. This is pure code motion plus the
visibility widening it requires; no behavior changes.

Two hot-path calls now cross a module boundary, so apply the perf skill's
layout rule to them: inline `streaming_template_already_sent` (reached from
`emit_css_module` once per rendered component on the ordinary path) and
`ensure_no_pending_streaming_root` (runs unconditionally at the end of every
fragment list), pushing the latter's two failure shapes into a `#[cold]`
`#[inline(never)]` constructor.

Also move `checkpoint_tag_buffers_retain_capacity_across_commits` next to the
inventory code it white-box tests, rather than reaching into private fields
from `lib.rs`.

Back-to-back Criterion A/B on `streaming_hydration_bench`:

| Case | Delta | Criterion verdict |
| --- | --- | --- |
| ordinary_render | +0.6% | No change detected |
| streaming B=1 | -3.4% | Improved |
| streaming B=3 | +1.8% | Within noise |
| streaming B=10 | -3.6% | Improved |
| streaming B=100 | -0.6% | Within noise |
| legacy one-shot 1/3/10/100 | -1.7% to -5.7% | Improved |

Ordinary render stays inside the 2% gate. 355 handler tests pass and
`cargo xtask check` is clean.
